+++ object/Cliente.js
@@ -1,0 +1,6 @@
+/* Definicion de clases */
+class Cliente {
+  nombreCliente;
+  dniCliente;
+  rutCliente;
+}
