+++ object/index.js
@@ -1,41 +1,8 @@
-/* Definicion de clases */
-class Cliente {
-  nombreCliente;
-  dniCliente;
-  rutCliente;
-}
-
-class CuentaCorriente {
-  numero;
-  #saldo;
-  agencia;
-
-  constructor() {
-    this.numero = '';
-    this.#saldo = 0;
-    this.agencia = '';
-  }
-
-  depositoEnCuenta(valor) {
-    if (valor > 0)
-      this.#saldo += valor;
-    return this.#saldo;
-  }
-
-  retirarDeCuenta(valor) {
-    if (valor <= this.#saldo)
-      this.#saldo -= valor;
-    return this.#saldo;
-  }
-
-  verSaldo() {
-    return this.#saldo;
-  }
-}
+/* import class */
+import { CuentaCorriente } from './CuentaCorriente.js';
 
 
-
-cuentaDeLeonardo = new CuentaCorriente();
+const cuentaDeLeonardo = new CuentaCorriente();
 // cuentaDeLeonardo.saldo = 0;
 
 let saldo = cuentaDeLeonardo.verSaldo();
