+++ object/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "cuentasbancopopular",
+  "version": "1.0.0",
+  "description": "Gestion de Cuentas Corrientes del Banco",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "Curso Alura",
+  "license": "ISC",
+  "type": "module"
+}
